refactor(ui): Overhaul UI system for modularity and maintainability
This commit introduces a major refactoring of the UI management system to improve its structure, maintainability, and scalability. The monolithic `uiManager.js` has been broken down into several smaller, more focused modules, and a clear separation of concerns has been established.

Key changes include:

- **Modularization:** The UI logic is now split into five distinct files located in `AddonExeBP/scripts/core/ui/`:
    - `panelRegistry.js`: Centralizes all UI panel definitions.
    - `panelBuilder.js`: Handles the logic for constructing UI forms.
    - `panelHandlers.js`: Manages the logic for processing form responses.
    - `actionRegistry.js`: Decouples UI actions from the core UI logic, centralizing all button-triggered functions.
    - `components.js`: Contains reusable UI components, such as the new generic confirmation dialog.

- **Decoupling:** An action registry (`actionRegistry.js`) has been implemented to serve as a single source for all UI-triggered functions. This removes direct dependencies from the UI code, making it cleaner and easier to manage.

- **Callback System for Resets:** A callback system has been implemented in `configManager.js` and `configurations.js`. When a configuration is reset, this system now triggers the appropriate re-initialization functions for the affected modules, ensuring that changes are applied to the live game state immediately without requiring a server reload.

- **Code Quality:** All related files have been cleaned up, removing unused imports and commented-out code. The linter now passes without errors, ensuring consistent code style.
