+++ AddonExeBP/scripts/core/configManager.js
@@ -1,5 +1,4 @@
 import createConfigManager from './configManagerFactory.js';
-import { configResetRegistry } from './configurations.js';
 import { deepClone } from './objectUtils.js';
 
 const mainConfigManager = createConfigManager('exe:config:current', '../config.js', 'Main', 'config');
@@ -13,26 +12,51 @@ export const updateMultipleConfig = mainConfigManager.updateMultiple;
 /**
  * Resets a section of the configuration to its default values.
  * @param {string} sectionKey The key of the config section to reset (e.g., 'tpa', 'homes'). Use 'all' to reset everything.
+ * @param {import('@minecraft/server').Player} [player] - The player who initiated the reset, for feedback.
  * @returns {Promise<{success: boolean, message: string}>}
  */
-export async function resetConfigSection(sectionKey) {
+export async function resetConfigSection(sectionKey, player) {
+    // Dynamically import configurations to break the circular dependency at load time.
+    const { configResetRegistry, configResetCallbacks } = await import('./configurations.js');
+
     if (sectionKey === 'all') {
         const resetPromises = [mainConfigManager.reset()];
         Object.values(configResetRegistry).forEach(config => resetPromises.push(config.reset()));
         await Promise.all(resetPromises);
-        return { success: true, message: 'All configuration settings have been reset to default.' };
+
+        // Trigger all post-reset callbacks
+        for (const key in configResetCallbacks) {
+            configResetCallbacks[key](player);
+        }
+        for (const key in configResetRegistry) {
+            if (configResetRegistry[key].postResetCallback) {
+                configResetRegistry[key].postResetCallback(player);
+            }
+        }
+
+        return { success: true, message: 'All configuration settings have been reset to default and systems reloaded.' };
     }
 
     if (configResetRegistry[sectionKey]) {
         await configResetRegistry[sectionKey].reset();
-        return { success: true, message: configResetRegistry[sectionKey].message };
+        if (configResetRegistry[sectionKey].postResetCallback) {
+            configResetRegistry[sectionKey].postResetCallback(player);
+        }
+        return { success: true, message: `${configResetRegistry[sectionKey].message} and reloaded.` };
     }
 
     // Dynamically import the latest default config to compare against
     try {
-        const { config: defaultConfig } = await import(`../config.js?v=${new Date().getTime()}`);
+        const { config: defaultConfig } = await import('../config.js');
         if (Object.prototype.hasOwnProperty.call(defaultConfig, sectionKey)) {
             updateConfig(sectionKey, deepClone(defaultConfig[sectionKey]));
+
+            // After resetting, check if there's a callback to re-initialize the system
+            if (configResetCallbacks[sectionKey]) {
+                configResetCallbacks[sectionKey](player);
+                return { success: true, message: `The '${sectionKey}' configuration has been reset and the system reloaded.` };
+            }
+
             return { success: true, message: `The '${sectionKey}' configuration section has been reset to default.` };
         } else {
             return { success: false, message: `Configuration section '${sectionKey}' not found.` };
