+++ AddonExeBP/scripts/core/configurations.js
@@ -1,4 +1,10 @@
 import createConfigManager from './configManagerFactory.js';
+import { restartAnnouncer } from '../modules/commands/announcement.js';
+import { initializeSpawnProtection } from '../modules/detections/spawnProtection.js';
+import { reloadRanks } from './rankManager.js';
+import { setLockState } from './playerDataManager.js';
+import { getConfig } from './configManager.js';
+
 
 const kitsConfigManager = createConfigManager('exe:kitsConfig:current', './kitsConfig.js', 'Kits', 'kitsConfig');
 const shopConfigManager = createConfigManager('exe:shopConfig:current', './shopConfig.js', 'Shop', 'shopConfig');
@@ -31,20 +37,56 @@ export const configResetRegistry = {
     'kits': {
         reset: resetKitsConfig,
         message: 'The \'kits\' configuration section has been reset to default.'
+        // No post-reset callback needed, data is read live.
     },
     'shop': {
         reset: resetShopConfig,
         message: 'The \'shop\' configuration section has been reset to default.'
+        // No post-reset callback needed, data is read live.
     },
     'spawn': {
         reset: resetSpawnConfig,
-        message: 'The \'spawn\' configuration section has been reset to default.'
+        message: 'The \'spawn\' configuration section has been reset to default.',
+        postResetCallback: (player) => {
+            initializeSpawnProtection();
+            if (player) {
+                player.sendMessage('§aSpawn protection system has been updated based on new settings.');
+            }
+        }
     },
     'ranks': {
         reset: resetRanksConfig,
-        message: 'The \'ranks\' configuration section has been reset to default.'
+        message: 'The \'ranks\' configuration section has been reset to default.',
+        postResetCallback: (player) => {
+            reloadRanks();
+            if (player) {
+                player.sendMessage('§aRanks have been reloaded with new settings.');
+            }
+        }
     }
 };
+
+/**
+ * A registry of functions to call after a specific config section is reset.
+ * This is for sections within the main `config.js` file.
+ */
+export const configResetCallbacks = {
+    'announcements': (player) => {
+        restartAnnouncer();
+        if (player) {
+            player.sendMessage('§aAnnouncement system has been updated with new settings.');
+        }
+    },
+    'dimensionLock': (player) => {
+        const config = getConfig(); // Get the freshly reset config
+        setLockState('nether', !!config.dimensionLock.lockNether);
+        setLockState('end', !!config.dimensionLock.lockEnd);
+        if (player) {
+            player.sendMessage('§aLive dimension lock states have been updated to match config.');
+        }
+    }
+};
+
 
 /**
  * Reloads all configurations that support it.
