+++ AddonExeBP/scripts/core/ui/actionRegistry.js
@@ -1,0 +1,343 @@
+import { world } from '@minecraft/server';
+import { ActionFormData, ModalFormData } from '@minecraft/server-ui';
+import { getPlayer } from '../playerDataManager.js';
+import * as utils from '../utils.js';
+import * as rulesManager from '../rulesManager.js';
+import * as helpfulLinksManager from '../helpfulLinksManager.js';
+import * as reportManager from '../reportManager.js';
+import * as bountyManager from '../bountyManager.js';
+import * as economyManager from '../economyManager.js';
+import * as tpaManager from '../tpaManager.js';
+import * as playerCache from '../playerCache.js';
+import { kickPlayer } from '../../modules/commands/kick.js';
+import { mutePlayer, unmutePlayer } from '../../modules/commands/mute.js';
+import { banPlayer, offlineBanPlayer, unbanPlayer } from '../../modules/commands/ban.js';
+import { freezePlayer, unfreezePlayer } from '../../modules/commands/freeze.js';
+import { showPanel } from '../uiManager.js';
+import { getConfig } from '../configManager.js';
+
+export const uiActionFunctions = {
+    showRules: async (player) => {
+        const rules = rulesManager.getRules();
+        const pData = getPlayer(player.id);
+
+        const rulesForm = new ActionFormData()
+            .title('§l§6Server Rules')
+            .body(rules.join('\n'));
+
+        if (pData && pData.permissionLevel <= 1) {
+            rulesForm.button('§l§4Edit Rules', 'textures/ui/icon_setting');
+        }
+
+        rulesForm.button('§l§8Close', 'textures/ui/cancel');
+
+        const response = await utils.uiWait(player, rulesForm);
+
+        if (response.canceled) { return; }
+
+        if (pData && pData.permissionLevel <= 1 && response.selection === 0) {
+            return showPanel(player, 'rulesManagementPanel');
+        }
+    },
+
+    showHelpfulLinks: async (player) => {
+        const links = helpfulLinksManager.getHelpfulLinks();
+        const pData = getPlayer(player.id);
+
+        const form = new ActionFormData()
+            .title('§l§9Helpful Links');
+
+        if (links.length === 0) {
+            form.body('§cNo helpful links have been configured by the admin.');
+        } else {
+            const bodyText = links.map(link => `§f${link.title}: §r${link.url}`).join('\n\n');
+            form.body(bodyText);
+        }
+
+        if (pData && pData.permissionLevel <= 1) {
+            form.button('§l§4Edit Links', 'textures/ui/icon_setting');
+        }
+
+        form.button('§l§8Close', 'textures/ui/cancel');
+
+        const response = await utils.uiWait(player, form);
+
+        if (response.canceled) { return; }
+
+        if (pData && pData.permissionLevel <= 1 && response.selection === 0) {
+            return showPanel(player, 'helpfulLinksManagementPanel');
+        }
+    },
+
+    assignReport: (player, context, panelId) => {
+        reportManager.assignReport(context.targetReport.id, player.id);
+        player.sendMessage(`§2Report ${context.targetReport.id} has been assigned to you.`);
+        showPanel(player, panelId, context);
+    },
+
+    resolveReport: (player, context) => {
+        reportManager.resolveReport(context.targetReport.id);
+        player.sendMessage(`§2Report ${context.targetReport.id} has been marked as resolved.`);
+        showPanel(player, 'reportListPanel');
+    },
+
+    clearReport: (player, context) => {
+        reportManager.clearReport(context.targetReport.id);
+        player.sendMessage(`§2Report ${context.targetReport.id} has been cleared.`);
+        showPanel(player, 'reportListPanel');
+    },
+
+    showUnbanForm: async (player) => {
+        const form = new ModalFormData().title('Unban Player').textField('Player Name', 'Enter player name');
+        const response = await utils.uiWait(player, form);
+        if (!response || response.canceled) { return true; }
+        const [targetName] = response.formValues;
+        if (!targetName) {
+            player.sendMessage('§cYou must enter a player name.');
+            return true;
+        }
+        unbanPlayer(player, targetName);
+        return true;
+    },
+
+    showUnmuteForm: async (player) => {
+        const form = new ModalFormData().title('Unmute Player').textField('Player Name', 'Enter player name');
+        const response = await utils.uiWait(player, form);
+        if (!response || response.canceled) { return true; }
+        const [targetName] = response.formValues;
+        if (!targetName) {
+            player.sendMessage('§cYou must enter a player name.');
+            return true;
+        }
+        unmutePlayer(player, targetName);
+        return true;
+    },
+
+    removeBounty: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const existingBounty = bountyManager.getBounty(targetPlayerId);
+
+        if (!existingBounty) {
+            player.sendMessage(`§c${targetPlayerName} does not have an active bounty.`);
+            return true;
+        }
+
+        bountyManager.removeBounty(targetPlayerId);
+        player.sendMessage(`§2Successfully removed the bounty from ${targetPlayerName}.`);
+        world.sendMessage(`§2The bounty on ${targetPlayerName} has been removed!`);
+
+        return true;
+    },
+
+    kickPlayer: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const targetPlayer = playerCache.getPlayerFromCache(targetPlayerId);
+        if (!targetPlayer) {
+            player.sendMessage(`§c${targetPlayerName} is not online.`);
+            return true;
+        }
+        const form = new ModalFormData().title(`Kick ${targetPlayerName}`).textField('Reason', 'Enter reason for kicking', { defaultValue: 'No reason provided.' });
+        const response = await utils.uiWait(player, form);
+        if (response && !response.canceled) {
+            const [reason] = response.formValues;
+            kickPlayer(player, targetPlayer, reason);
+        }
+        return true;
+    },
+
+    freezePlayer: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const targetPlayer = playerCache.getPlayerFromCache(targetPlayerId);
+        if (!targetPlayer) {
+            player.sendMessage(`§c${targetPlayerName} is not online.`);
+            return true;
+        }
+        freezePlayer(player, targetPlayer);
+        return true;
+    },
+
+    unfreezePlayer: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const targetPlayer = playerCache.getPlayerFromCache(targetPlayerId);
+        if (!targetPlayer) {
+            player.sendMessage(`§c${targetPlayerName} is not online.`);
+            return true;
+        }
+        unfreezePlayer(player, targetPlayer);
+        return true;
+    },
+
+    unmutePlayer: async (player, context) => {
+        const { targetPlayerName } = context;
+        unmutePlayer(player, targetPlayerName);
+        return true;
+    },
+
+    mutePlayer: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const targetPlayer = playerCache.getPlayerFromCache(targetPlayerId);
+        if (!targetPlayer) {
+            player.sendMessage(`§c${targetPlayerName} is not online. Use /offlinemute instead.`);
+            return true;
+        }
+        const form = new ModalFormData().title(`Mute ${targetPlayerName}`).textField('Duration', 'e.g., 30m, 2h, 7d. Default: perm', { defaultValue: 'perm' }).textField('Reason', 'Enter reason for muting', { defaultValue: 'No reason provided.' });
+        const response = await utils.uiWait(player, form);
+        if (response && !response.canceled) {
+            const [duration, reason] = response.formValues;
+            mutePlayer(player, targetPlayer, duration, reason);
+        }
+        return true;
+    },
+
+    banPlayer: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const form = new ModalFormData().title(`Ban ${targetPlayerName}`).textField('Duration', 'e.g., 30m, 2h, 7d. Default: perm', { defaultValue: 'perm' }).textField('Reason', 'Enter reason for banning', { defaultValue: 'No reason provided.' });
+        const response = await utils.uiWait(player, form);
+        if (response && !response.canceled) {
+            const [duration, reason] = response.formValues;
+            const targetPlayer = playerCache.getPlayerFromCache(targetPlayerId);
+            if (targetPlayer) {
+                banPlayer(player, targetPlayer, duration, reason);
+            } else {
+                offlineBanPlayer(player, targetPlayerId, targetPlayerName, duration, reason);
+            }
+        }
+        return true;
+    },
+
+    tpaPlayer: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const targetPlayer = playerCache.getPlayerFromCache(targetPlayerId);
+        if (!targetPlayer) {
+            player.sendMessage(`§c${targetPlayerName} is not online.`);
+            return true;
+        }
+        if (player.id === targetPlayer.id) {
+            player.sendMessage('§cYou cannot send a TPA request to yourself.');
+            return true;
+        }
+        const result = tpaManager.createRequest(player, targetPlayer, 'tpa');
+        if (result.success) {
+            player.sendMessage(`§2TPA request sent to ${targetPlayerName}.`);
+            targetPlayer.sendMessage(`§2${player.name} has requested to teleport to you. Use !tpaccept or !tpadeny.`);
+        } else {
+            player.sendMessage(`§cError: ${result.message}`);
+        }
+        return true;
+    },
+
+    tpaherePlayer: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const targetPlayer = playerCache.getPlayerFromCache(targetPlayerId);
+        if (!targetPlayer) {
+            player.sendMessage(`§c${targetPlayerName} is not online.`);
+            return true;
+        }
+        if (player.id === targetPlayer.id) {
+            player.sendMessage('§cYou cannot send a TPAHere request to yourself.');
+            return true;
+        }
+        const result = tpaManager.createRequest(player, targetPlayer, 'tpahere');
+        if (result.success) {
+            player.sendMessage(`§2TPAHere request sent to ${targetPlayerName}.`);
+            targetPlayer.sendMessage(`§2${player.name} has requested for you to teleport to them. Use !tpaccept or !tpadeny.`);
+        } else {
+            player.sendMessage(`§cError: ${result.message}`);
+        }
+        return true;
+    },
+
+    bountyPlayer: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const form = new ModalFormData().title(`Set Bounty on ${targetPlayerName}`).textField('Amount', 'Enter amount');
+        const response = await utils.uiWait(player, form);
+        if (response && !response.canceled) {
+            const [amountStr] = response.formValues;
+            const amount = Number(amountStr);
+            const config = getConfig();
+            if (isNaN(amount) || amount < config.bounties.minimumBounty) {
+                player.sendMessage(`§cInvalid amount. The minimum bounty is $${config.bounties.minimumBounty}.`);
+                return true;
+            }
+            if (economyManager.getBalance(player.id) < amount) {
+                player.sendMessage('§cYou do not have enough money for this bounty.');
+                return true;
+            }
+
+            const result = economyManager.removeBalance(player.id, amount);
+            if (result) {
+                bountyManager.incrementBounty(targetPlayerId, amount);
+                player.sendMessage(`§2You have placed a bounty of §6$${amount}§2 on ${targetPlayerName}.`);
+                world.sendMessage(`§cSomeone has placed a bounty of §6$${amount}§c on ${targetPlayerName}!`);
+            } else {
+                player.sendMessage('§cFailed to place bounty.');
+            }
+        }
+        return true;
+    },
+
+    reportPlayer: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const form = new ModalFormData().title(`Report ${targetPlayerName}`).textField('Reason for report:', 'Enter the reason here');
+        const response = await utils.uiWait(player, form);
+        if (response.canceled) {
+            player.sendMessage('§cReport canceled.');
+            return true;
+        }
+        const [reason] = response.formValues;
+        if (!reason || reason.trim().length === 0) {
+            player.sendMessage('§cYou must provide a reason.');
+            return true;
+        }
+        reportManager.createReport(player, targetPlayerId, targetPlayerName, reason);
+        player.sendMessage('§2Report submitted. Thank you for your help.');
+        return true;
+    },
+
+    removePlayerBounty: async (player, context) => {
+        const { targetPlayerId, targetPlayerName } = context;
+        const targetBounty = bountyManager.getBounty(targetPlayerId);
+
+        if (!targetBounty) {
+            player.sendMessage(`§c${targetPlayerName} does not have an active bounty.`);
+            return true;
+        }
+
+        const form = new ModalFormData()
+            .title(`Remove Bounty from ${targetPlayerName}`)
+            .textField(`Bounty Amount: $${targetBounty.amount.toFixed(2)}\nEnter amount to remove:`, 'Enter amount');
+
+        const response = await utils.uiWait(player, form);
+
+        if (response && !response.canceled) {
+            const [amountStr] = response.formValues;
+            const amount = Number(amountStr);
+
+            if (isNaN(amount) || amount <= 0) {
+                player.sendMessage('§cInvalid amount. Please enter a positive number.');
+                return true;
+            }
+
+            if (amount > targetBounty.amount) {
+                player.sendMessage(`§cYou cannot remove more than the bounty amount ($${targetBounty.amount.toFixed(2)}).`);
+                return true;
+            }
+
+            if (economyManager.getBalance(player.id) < amount) {
+                player.sendMessage('§cYou dont have enough money for this!');
+                return true;
+            }
+
+            const result = economyManager.removeBalance(player.id, amount);
+            if (result) {
+                bountyManager.incrementBounty(targetPlayerId, -amount);
+                player.sendMessage(`§2You have removed $${amount.toFixed(2)} from ${targetPlayerName}'s bounty.`);
+                world.sendMessage(`§2${player.name} has removed $${amount.toFixed(2)} from ${targetPlayerName}'s bounty!`);
+            } else {
+                player.sendMessage('§cFailed to remove bounty.');
+            }
+        }
+
+        return true;
+    }
+};
